Fixed a bug that caused bitcoind to crash; some enteries in /proc/stat are too big for stoi; use stol instead

+++ src/logFile.cpp
@@ -382,7 +382,7 @@ long getTotalTime()
     ss >> word;
     long total = 0;
     while(ss >> word)
-        total += std::stoi(word);
+        total += std::stol(word);
 
     return total;
 }
